requirements.txt
titiler.core==0.7.0
rio-tiler-mvt==0.0.1dev2
starlette-cramjam>=0.3,<0.4
mangum>=0.10.0
